Conflicts:
	client/src/main/resources/jorm.properties
	client/src/main/webapp/WEB-INF/vsl/player/state.vsl
	utils/src/main/java/com/distributed/springtest/utils/records/Generator.java

+++ gamecontent/src/main/java/com/distributed/springtest/gamecontent/GameContentController.java
@@ -39,80 +39,64 @@ public class GameContentController {
         return new ResponseEntity<List<BuildingCost>>(buildingCosts, HttpStatus.OK);
     }
 
-    @RequestMapping("/buildings/{id}/costs/add")
+    @RequestMapping(value = "/buildings/{id}/costs/add", method = RequestMethod.POST)
     public ResponseEntity<String> addBuildingCost(@PathVariable Integer id, @RequestBody BuildingCost cost) throws SQLException {
         BuildingCost buildingCost = BuildingCost.select(BuildingCost.class, "SELECT * FROM building_costs WHERE building_id = #1# AND resource_id = #2#", cost.getBuildingId(), cost.getResourceId());
         if(buildingCost == null) {
-            cost.setBuildingId(id);
-            cost.save();
-            cost.transaction().commit();
+            buildingCost = new BuildingCost();
+            buildingCost.setBuildingId(id);
+            buildingCost.setResourceId(cost.getResourceId());
+            buildingCost.setAmount(cost.getAmount());
+            buildingCost.save();
+            buildingCost.transaction().commit();
+            return new ResponseEntity<String>("OK", HttpStatus.OK);
         } else {
+            return new ResponseEntity<String>("A cost with that resource type already exists", HttpStatus.BAD_REQUEST);
+        }
+    }
+
+    @RequestMapping(value = "/buildings/{id}/costs/{costId}/modify", method = RequestMethod.PUT)
+    public void modifyBuildingCost(@PathVariable Integer id, @PathVariable Integer costId, @RequestBody BuildingCost cost) throws SQLException {
+        BuildingCost buildingCost = BuildingCost.findById(BuildingCost.class, cost.getId());
+        if(buildingCost != null && buildingCost.getBuildingId().equals(id)) {
+            buildingCost.setResourceId(cost.getResourceId());
             buildingCost.setAmount(cost.getAmount());
             buildingCost.save();
             buildingCost.transaction().commit();
         }
-        return new ResponseEntity<String>("OK", HttpStatus.OK);
     }
 
-    @RequestMapping("/buildings/{id}/costs/delete")
-    public ResponseEntity<String> deleteBuildingCost(@PathVariable Integer id, @RequestBody BuildingCost cost) throws SQLException {
-        BuildingCost buildingCost  = BuildingCost.select(BuildingCost.class, "SELECT * FROM building_costs WHERE building_id = #1# AND resource_id = #2#", cost.getBuildingId(), cost.getResourceId());
-        if(buildingCost != null) {
+    @RequestMapping(value = "/buildings/{id}/costs/{costId}/delete", method = RequestMethod.DELETE)
+    public void deleteBuildingCost(@PathVariable Integer id, @PathVariable Integer costId) throws SQLException {
+        BuildingCost buildingCost  = BuildingCost.findById(BuildingCost.class, costId);
+        if(buildingCost != null && buildingCost.getBuildingId().equals(id)) {
             buildingCost.delete();
-            return new ResponseEntity<String>("OK", HttpStatus.OK);
-        } else {
-            return new ResponseEntity<String>("No such object", HttpStatus.BAD_REQUEST);
+            buildingCost.transaction().commit();
         }
     }
 
     @RequestMapping(value = "/buildings/add", method = RequestMethod.POST)
-    public Integer addBuilding(@RequestBody BuildingInfoWrapper buildingInfoWrapper) throws SQLException {
-        BuildingInfo incomingBuildingInfo = buildingInfoWrapper.getBuildingInfo();
+    public ResponseEntity<Integer> addBuilding(@RequestBody BuildingInfo incomingBuildingInfo) throws SQLException {
         BuildingInfo buildingInfo = new BuildingInfo();
         buildingInfo.setName(incomingBuildingInfo.getName());
         buildingInfo.setGeneratedId(incomingBuildingInfo.getGeneratedId());
         buildingInfo.setGeneratedAmount(incomingBuildingInfo.getGeneratedAmount());
+        buildingInfo.setBuildtime(incomingBuildingInfo.getBuildtime());
         buildingInfo.save();
-        List<BuildingCost> buildingCosts = buildingInfoWrapper.getBuildingCosts();
-        for(BuildingCost incomingBuildingCost : buildingCosts) {
-            BuildingCost buildingCost = new BuildingCost();
-            buildingCost.setResourceId(incomingBuildingCost.getResourceId());
-            buildingCost.setBuildingId(incomingBuildingCost.getBuildingId());
-            buildingCost.setAmount(incomingBuildingCost.getAmount());
-            buildingCost.save();
-        }
         buildingInfo.transaction().commit();
-        return buildingInfo.getId();
+        return new ResponseEntity<Integer>(buildingInfo.getId(), HttpStatus.OK);
     }
 
     @RequestMapping(value = "/buildings/modify", method = RequestMethod.PUT)
-    public Integer modifyBuilding(@RequestBody BuildingInfoWrapper buildingInfoWrapper) throws SQLException {
-        BuildingInfo incomingBuildingInfo = buildingInfoWrapper.getBuildingInfo();
+    public void modifyBuilding(@RequestBody BuildingInfo incomingBuildingInfo) throws SQLException {
         BuildingInfo buildingInfo = Building.findById(BuildingInfo.class, incomingBuildingInfo.getId());
-        buildingInfo.setName(incomingBuildingInfo.getName());
-        buildingInfo.setGeneratedAmount(incomingBuildingInfo.getGeneratedAmount());
-        buildingInfo.setGeneratedId(incomingBuildingInfo.getGeneratedId());
-        buildingInfo.save();
-        List<BuildingCost> buildingCosts = buildingInfoWrapper.getBuildingCosts();
-        List<BuildingCost> currentBuildingCosts = BuildingCost.selectAll(BuildingCost.class,
-                "SELECT * FROM building_costs WHERE building_id = #1#", buildingInfo.getId());
-        List<BuildingCost> buildingCostsForRemoval = new ArrayList<BuildingCost>(currentBuildingCosts);
-        buildingCostsForRemoval.removeAll(buildingCosts);
-        for(BuildingCost buildingCost : buildingCostsForRemoval) {
-            buildingCost.delete();
+        if(buildingInfo != null) {
+            buildingInfo.setName(incomingBuildingInfo.getName());
+            buildingInfo.setGeneratedAmount(incomingBuildingInfo.getGeneratedAmount());
+            buildingInfo.setGeneratedId(incomingBuildingInfo.getGeneratedId());
+            buildingInfo.save();
+            buildingInfo.transaction().commit();
         }
-        for(BuildingCost incomingBuildingCost : buildingCosts) {
-            BuildingCost buildingCost = BuildingCost.select(BuildingCost.class, "SELECT * FROM building_costs WHERE building_id = #1# AND resource_id = #2#", buildingInfoWrapper.getBuildingInfo().getId(), incomingBuildingCost.getResourceId());
-            if(buildingCost == null) {
-                buildingCost = new BuildingCost();
-            }
-            buildingCost.setResourceId(incomingBuildingCost.getResourceId());
-            buildingCost.setBuildingId(buildingInfoWrapper.getBuildingInfo().getId());
-            buildingCost.setAmount(incomingBuildingCost.getAmount());
-            buildingCost.save();
-        }
-        buildingInfo.transaction().commit();
-        return buildingInfo.getId();
     }
 
     @RequestMapping(value = "/resources/add", method = RequestMethod.POST)
